devel/binutils: Add devel/elfutils dependency options
Currently binutils auto-discovers libdebuginfod.so existence in
$LOCALBASE and if found, it will use it. This doesn't matter when
building under poudriere because devel/elfutils is never
pre-installed as a dependency.

However users building binutils on systems with a pre-installed
devel/elfutils will discover that binutils will auto-detect the
existence of libdebuginfod.so and use it. This is not always desired.
Therefore add an option so that users who build binuitls by hand
can optionally force binutils not to detect or use libdebuginfod.so.

On the flip side other users may want to depend on devel/elfutils and
have the ports infrastructure preinstall it for them.

The default option is to auto-detect and depend on devel/elfutils if
found.

PR:		273798