devel/boost-{all,libs}: Adjust range of OSVERSION to use LLVM from ports
The reason to use LLVM from ports is that there is bug in Clang/LLVM
of base system that causes assertion failure while building
devel/boost-libs. The bug is already fixed with main, stable/14 and
stable/13 branches of src repository. So we need to use LLVM from
ports only with following ranges of OSVERSION.

1. 1500000 <= OSVERSION < 1500017
2. 1400000 <= OSVERSION < 1400509
3. 1300000 <= OSVERSION < 1303502

And commit 3b8c51bd37b0 adjusts the range as above.

But there is a pitfall about it. The assertion failure happens only if
base system is built with WITH_LLVM_ASSERTIONS variable set. By
default it is set with main branch but not with stable/* and releng/*
ones. So wheareas the failure happens with -CURRENT even if it is
built with default variables settings, it doesn't happen with -STABLE
and -RELEASE unless they are built with WITH_LLVM_ASSERTIONS set. IMHO
majority of FreeBSD users use -RELEASE with official binary, and few
-STABLE users set WITH_LLVM_ASSERTIONS. As a result adjustment with
commit 3b8c51bd37b0 forces most -RELEASE and -STABLE users to build
LLVM port unnecessarily.

So adjust the range so LLVM from ports is used by default only if base
system is -CURRENT with the assertion failure unfixed. And as a last
resort for -STABLE and -RELEASE users who set WITH_LLVM_ASSERTIONS,
add LLVM_FROM_PORTS option to devel/boost-libs that forces to use LLVM
from ports.

PR:		277856
Approved by:	maintainer timeout