mail/rubygem-extended_email_reply_parser: Add rubygem-extended_email_reply_parser 0.5.1
When implementing a "reply or comment by email" feature, it's necessary to
filter out signatures and the previous conversation. One needs to extract just
the relevant parts for the conversation or comment section of the application.
This is what this ruby gem helps to do.

This gem is an extended version of GitHub's email_reply_parser. It wraps the
original email_reply_parser and allows to build extensions such as support for
i18n and detecting previous conversation that is not properly marked as
quotation by the sender's mail client.