x11/sddm: Update to 0.21.0
Sticking with Qt5 for now. The Qt6 version basically works, i.e., it
starts up and one can effectively log in, but it doesn't seem to respect
the user-selected theme. More investigation is required to this point.

The requirement of ECM has thankfully been dropped.

Disable the qtvirtualkeyboard InputMethod by default as it doesn't seem
to work properly under Qt5 for any included theme, even in the previous
version, displaying only a black box covering half of the screen where
the virtual keyboard should be. The breeze theme (installed by
x11/plasma(5|6)-plasma-workspace) supresses this behavior, and is IMHO
the most intuitive theme to use, but when the 'virtual keyboard' button
is clicked, a black box still appears. This change removes that button,
elimitating further user confusion. Even when working properly under Qt6,
it is an annoyance, at best, to be bombarded by a virtual keyboard when
a physical keyboard is present. My impression is, the vast majority of
FreeBSD users are likely to be using a physical keyboard by default and
don't want to be bothered by clicking down a functional virtual keyboard
or to be confused by a black box taking up half of the screen.

Patch self-installed files using the ports framework instead of using
eyebrow-raising incantations that lead to staging errors. It's generally
not a good idea to use REINPLACE_CMD on STAGEDIR as it leaves unwanted
*.bak files.

https://github.com/sddm/sddm/releases/tag/v0.21.0