net/wifibox-alpine: Move away from the use of Alpine Linux mirrors
Alpine Linux packages are maintained in a rolling fashion and
periodic snapshots seem to be gone recently.  Store the specific
APK versions of dependencies on the project's GitHub to unbreak the
build when optional packages (e.g. mDNSResponder) are configured to
use and which are not cached by the FreeBSD distfile mirrors.

While here, start mirroring the root file system tarball in use,
just to become completely independent from the Alpine Linux CDN.