x11-fonts/geist-mono: Add geist-mono 1.3.0
Geist is a new font family for Vercel, created by Vercel in collaboration with
Basement Studio.

Geist Mono is a monospaced typeface that has been crafted to be the perfect
partner to Geist Sans. It is designed to be used in code editors, diagrams,
terminals, and other textbased interfaces where code is represented.