x11-fonts/geist-sans: Add geist-sans 1.3.0
Geist is a new font family for Vercel, created by Vercel in collaboration with
Basement Studio.

Geist Sans is a sans-serif typeface designed for legibility and simplicity. It
is a modern, geometric typeface that is based on the principles of classic Swiss
typography. It is designed to be used in headlines, logos, posters, and other
large display sizes.