devel/llvm17: install all heaaders with clang
Historically we've been unable to build FreeBSD with the full set due to
conflicts and/or missing features between the compiler provided headers
and FreeBSD's headers. I've verified that I can build world and kernel
on the main, stable/14, and stable/13 branches for amd64 so let's give
it another try in broader testing.

PR:		274542