audio/cantata: Update to 3.2.0
Switch to active fork as the original author archived their repo and is
no longer working on it. [1]

The most notable changes are that the project has been ported to Qt6
and has added support for TagLib 2.x.

The QT5 option has been renamed to QTMULTIMEDIA. Enable QTMULTIMEDIA for
streaming support by default instead of VLC as vlc still uses Qt5 and
pulls in many more unneeded dependencies.

Disable audio CD support by default. It builds, but doesn't work, even
in the previous version. It will take some more work to port the udisks2
backend of the internal trimmed-down copy of Solid to bsdisks. This is
primarily an MPD frontend and optical media has generally fallen out of
popularity, so this is a bit of a rainy day project.

Pet portclippy(1) and portlint(1). Whitespace cleanup.

PR:		279462 [1]
Reported by:	diizzy [1]