net-im/libaccounts-qt: Fix build if DOCS is off
Move BINARY_ALIAS to the DOCS option so that the build doesn't break
due to qhelpgenerator not being installed.

Add graphics/graphviz as a build dependency for DOCS. If the user has
the GRAPHVIZ option off in devel/doxygen, it does not get installed and
causes pkg-plist errors.

PR:		279946
Reported by:	<tjlegg@gmail.com>