sysutils/fluent-bit: Remove non-functional EXAMPLES variables
The EXAMPLES option has never been introduced and so EXAMPLES_CMAKE_*
variables serve no purpose. It's better to remove them.

I tried adding the EXAMPLES option and tested the build both with the
option on and off. It looks like Fluent Bit does not install any of its
examples even when it builds them. Because of that the better fix is to
not have the EXAMPLES option at all for now.

Approved by:	portmgr blanket
Fixes:		228b05d23612 Upgrade to 1.4.2
Sponsored by:	trivago N.V.
Sponsored by:	Klara, Inc.