databases/mongodb80: update to 8.0.0-rc16
Changes:
https://github.com/mongodb/mongo/compare/r8.0.0-rc15...r8.0.0-rc16

Removed SSL from OPTIONS. It was inherited from the first versions
of databases/mongodb, long before I became maintainer.
Currently disabling the SSL option does not disable ssl as ssl is
auto-detected and disabling ssl needs explicit --ssl=off which is
not implemented in the port. I never had a request for this in years
so I decided to just remove the option and have ssl enabled always.