math/blis: fix build with clang 19
Clang 19 has removed support for the -mavx512pf and -mavx512er flags,
for the Intel Xeon Phi (aka "Knights Landing"/knl or "Knights
Mill"/knm) [1] [2].

This causes fatal errors during configuration of math/blis:

    Compiling obj/x86_64/kernels/knl/1m/bli_dpackm_knl_asm_24x8.o ('knl' CFLAGS for kernels)
    cc: fatal error: unknown argument '-mavx512pf'; did you mean '-mavx512f'?
    gmake: *** [Makefile:653: obj/x86_64/kernels/knl/1m/bli_dpackm_knl_asm_24x8.o] Error 1

Add an EXTRA_PATCHES entry to deal with this situation.

Note that in the future, this may also have to be done for gcc 15 and
later. In that case, the patch might be done unconditionally.

[1] https://github.com/llvm/llvm-project/pull/75580
[2] https://github.com/llvm/llvm-project/pull/92883

PR:		280783
Approved by:	maintainer timeout (2 weeks)
MFH:		2024Q3