textproc/R-cran-downlit: new port, Syntax Highlighting and Automatic Linking
Syntax highlighting of R code, specifically designed for the needs of
'RMarkdown' packages like 'pkgdown', 'hugodown', and 'bookdown'. It includes
linking of function calls to their documentation on the web, and automatic
translation of ANSI escapes in output to the equivalent HTML.

The math/R doesn't ship grid library's vignette so should remove one
test.

This port is needed to update of devel/R-cran-devtools.