devel/cmake-core: Refactor
Move the update reminder to Uses/Mk/cmake.mk where the version has
been changed since after 7bdeb4414c22 and will be more likely noticed.

Switch to an out-of-source build. This eliminates the need to remove
certain unwanted files prior to install.

Define a sensible CONFIGURE_LOG for users to report as the default
doesn't exist.

Sort CONFIGURE_ARGS, not necessarily alphabetically, but to improve
readability.

Pet portclippy(1): hoist the post-patch target.

Settle on one sed(1) delimiter instead of three. Pipes (|) seems to
be the most readable here.

Remove cruft from ${FILESDIR}/InitialCache.cmake and convert to SUB_FILES
to take advantage of framework substitutions. archivers/lzmalib is gone,
so there is no need to patch for that. Remove the CPACK_OPTION_COMMENT
variable in favor of CMake logic.