japanese/font-bizin-gothic: New port: Font for programming composed of BizUD gothic and Inconsolata
Bizin Gothic is a font for programming composing the universal
design font BizUD Gothic and the latin font Inconsolata.
The goal is to combine the eye-friendly glyphs of BIZ UD Gothic
and the quirky and beautiful glyphs of Inconsolata without any
sense of discomfort.
In addition, the specifications for various character
adjustments to improve readability in this font were made with
Ricty in mind.

PR:	279913
Reported by:	Hiroo Ono <hiroo.ono+freebsd@gmail.com>