sysutils/zfs-replicate: change maintainer to tschetter.victor@gmail.com
This port was last updated in September of 2022, and the maintainer seems AWOL.

Since then, features have been added to the original github repository (located at https://github.com/aaronhurt/zfs-replicate)

I have emailed the maintainer listed on the port page (jnlin@freebsd.cs.nctu.edu.tw) but the only response was that the message could not be delivered. This was 48 hours ago on August 19, 2024.

I would like to volunteer to maintain this port, as I am the one who pushed the changes to the github repository to begin with.

PR: 280986