math/labplot: Update to 2.11.1
Support everything that FreeBSD currently has to offer.
We don't have ports for Mqtt[5|6] nor for qt[5|6]advanceddocking
at time of commit.

Upstream seems to have had a problem with readstat on FreeBSD and
disabled it by default [1], but the issue has not manifested here.
Readstat support is now optional in case there is a problem that
needs to be investigated further.

Improve options descriptions, but keep them inline.

[1] https://invent.kde.org/education/labplot/-/commit/b2c4bf27105f56e25e116505a67173205d91eeb8

Latest announcement (2.11.0):
https://labplot.kde.org/2024/07/16/labplot-2-11-released/