textproc/docproj: Update gohugo dependency
Bump it to 0.133.1

Merge to quarterly in a couple of days.
After that, wait until the package is ready and commit some changes to the
website part of the documentation that would not build with the gohugo version
currently in Q3 (namely changing from Site.First to Site.Default since the
former is deprecated an will stop working).

Approved by:	doceng@ (implicit)
MFH:		2024Q3