net/wifi-firmware-kmod: add more MASTER_SITES
pkg builders are failing to fetch all firmware files due to
'Service Unavailable' errors.  We assume it is some kind of rate
limit on the git.kernel.org site and not something local to the
pkg builder setup.  I was unable to reproduce it locally.
Given git.kernel.org is a geolocated service unroll the server names
as well so that pkg builders can fall back to further machines [1].
We do leave the general name first so the common user should not
see any problems.

Suggested by:	dbaio [1]
Sponsored by:	The FreeBSD Foundation
Reviewed by:	jrm
Differential Revision: https://reviews.freebsd.org/D46681