devel/gitea-act_runner: New port
Act runner is a runner for Gitea based on Gitea fork of act.

This is a fork of the original act runner developed and maintained by Gitea
community and intended to be used with Gitea only.

PLEASE NOTE that this port still does not ship with a rc script as this is
a WIP at the moment.

WWW: https://gitea.com/gitea/act_runner/