Mk/bsd.sites.mk: don't add GH to MASTER_SITES when USE_GITHUB=nodefault
When using USE_GITHUB=nodefault, we don't need to add GH to
MASTER_SITES, as the github urls that are needed will be added later.

Adding GH here means we'll add a generic GitHub URL to the fetch list,
in the form of github.com/<portname>/<portname>/.../<dist> which will
not work. When used with USE_GITLAB the framework will add this bogus
URL before the correct gitlab url. It does not prevent the distfile to
be fetched from gitlab, but it makes everyone try and fetch from a bogus
url before trying the correct url.

PR:	279816